Extract Excel workbook helpers
